+++ changelog.md
@@ -5,10 +5,12 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ---
 
-## 0.1.0 - Unreleased
+## 0.1.0 - 2026-09-06
 ### Added
 - Header
 - Footer
 - Hero Section
 - About Section
 - Search Popup UI
+- Experiments Section
+- Templates Section